New!  Recursive Buildfiles.
The build tool, while not actually functional, now understands a system of
recursive files all named Buildfile, in various and sundry directories.  Also,
such a system now exists, though it isn't fully populated with information.

The semantics are that all the Buildfiles everywhere in the project tree are
loaded and parsed, and the targets they all provide are merged together into
one namespace.  Unlike in the GNU recursive Makefile template, which is
famously inefficient, this is done all within one process, and therefore
should scale well (or at least, not as badly).

Some semantics not yet implemented: The target to build may be specified on
the command line; if not, the default is always given by the Buildfile in
the current directory - for, YES!, this system can be invoked from
subdirectories, and will still do the right thing! - or, if it is not given
locally, an error is produced and nothing is built.  You probably don't want
to build the entire project from a subdirectory; if you do, it's easily
done, but why are you not in the project root?